### 2024-08-07 #7083 @editorialbot
[PRE REVIEW]: Mango.jl: A Julia-Based Multi-Agent Simulation Framework
**Submitting author:** @jsagerOffis (<a href="http://orcid.org/0000-0001-6352-4213">Jens Sager</a>)
**Repository:** https://github.com/OFFIS-DAI/Mango.jl
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.3.0
