### 2024-08-07 #7083 @danielskatz
👋 @fraukewiese & @jbytecode - I know both of you are currently at your editing limits, but I wanted to see if this might be of interest to either of you to edit (based on the energy/Julia aspects respectively); perhaps you are almost done with a current assignment...  If neither of you are able to handle it, which is perfectly fine, I will take it.
