### 2024-08-07 #7083 @jbytecode
@jsagerOffis - Hi, I am the handling editor.

Yes, you can change the branch of the paper. Please create the new branch and carry the manuscript and supplementary files into the new branch. Please ping me when you've done with it.
