### 2024-08-07 #7083 @jsagerOffis
@jbytecode The paper is on the `joss_paper` branch and the correct `0.3.0` state of the software is now also merged to that branch.

I have also added the julia reference to the manuscript.
