### 2024-08-07 #7083 @jsagerOffis
> @jsagerOffis - Have you performed a CrossRef search for the citations with missing DOIs?

Yes, I have added DOIs to all references where I could find one.
