### 2024-08-07 2c0f0ca
add installation instructions to readme
