### 2024-08-09 #7083 @jbytecode
👋👋👋 Dear @Libbum 👋👋👋

**[[PRE REVIEW]: Mango.jl: A Julia-Based Multi-Agent Simulation Framework](https://github.com/openjournals/joss-reviews/issues/7083)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/7083)).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be nearly 20 check items for each single reviewer.

Thank you in advance!
