### 2024-08-10 #7098 @Tortar
Hi @rcschrg @jsagerOffis !

I read the article and given that there is stated that

```
The main reason for this julia-based version is to allow better focus on simulation36
performance, enabling larger scales of multi-agent simulations
```

I think a good addition to the paper would be a benchmark between the Python and Julia version of the software. Do you already have some examples implemented in both frameworks to use for this?
