### 2024-08-12 #7098 @jsagerOffis
Hi @Tortar !

We have did some benchmarks like that for the proof of concept version of Mango.jl.
I had already planned to update those to the current version since a lot has changed since then, so I will get on that asap now.
