### 2024-08-12 #7098 @jsagerOffis
@Tortar I have just updated our old benchmark scripts to the new Mango.jl version and added a small performance section to the paper.

I have linked the benchmark repository in the paper and we should be able to set that one to be public as well shortly.
