### 2024-08-13 6734d3d
Implementing forwarding as convenience send method, and implement auto-forward rules.
