+++ src/agent/core.jl
@@ -8,12 +8,18 @@ export @agent,
     stop_and_wait_for_all_tasks,
     shutdown,
     on_ready,
-    on_start
+    on_start,
+    forward_to,
+    add_forwarding_rule,
+    delete_forwarding_rule,
+    ForwardingRule
 
 using UUIDs
 
 import Dates
 
+FORWARDED_FROM_ADDR = "forwarded_from_address"
+FORWARDED_FROM_ID = "forwarded_from_id"
 
 """
 Context of the agent. Represents the environment for the specific agent. Therefore it includes a 
@@ -35,6 +41,12 @@ struct AgentRoleHandler
     models::Dict{DataType,Any}
 end
 
+struct ForwardingRule
+    from_address::AgentAddress
+    to_address::AgentAddress
+    forward_replies::Bool
+end
+
 """
 All baseline fields added by the @agent macro are listed in this vector.
 They are added in the same order defined here.
@@ -46,6 +58,7 @@ AGENT_BASELINE_FIELDS::Vector = [
     :(scheduler::AbstractScheduler),
     :(aid::Union{Nothing,String}),
     :(transaction_handler::Dict{String,Tuple}),
+    :(forwarding_rules::Vector{ForwardingRule})
 ]
 
 """
@@ -60,6 +73,7 @@ AGENT_BASELINE_DEFAULTS::Vector = [
     () -> Scheduler(),
     () -> nothing,
     () -> Dict(),
+    () -> Vector()
 ]
 
 """
@@ -134,13 +148,38 @@ macro agent(struct_def)
     esc(Expr(:block, new_struct_def, default_constructor_def))
 end
 
+
+function build_forwarded_address_from_meta(meta::AbstractDict)
+    return AgentAddress(aid=meta["reply_to_forwarded_from_id"], address=meta["reply_to_forwarded_from_address"], tracking_id=get(meta, TRACKING_ID, nothing))
+end
+
 """
 Internal API used by the container to dispatch an incoming message to the agent. 
 In this function the message will be handed over to the different handlers in the
 agent.
 """
 function dispatch_message(agent::Agent, message::Any, meta::AbstractDict)
+    # check if auto forwarding is applicable
+    sender_addr = get(meta, SENDER_ADDR, nothing)
+    sender_id = get(meta, SENDER_ID, nothing)
+    forwarded = false
+    for rule::ForwardingRule in agent.forwarding_rules
+        if rule.from_address.aid == sender_id && rule.from_address.address == sender_addr
+            wait(forward_to(agent, message, rule.to_address, meta))
+            forwarded = true
+        end
+        # if reply to a forwarded message and replies shall be forwarded, forward this message to the original sender
+        if rule.to_address.address == sender_addr && rule.to_address.aid == sender_id && get(meta, "reply_to_forwarded", false) && rule.forward_replies
+            wait(forward_to(agent, message, build_forwarded_address_from_meta(meta), meta))
+            forwarded = true
+        end
+    end
+    if forwarded
+        return
+    end
+    
     lock(agent.lock) do
+        # check if part of a transaction
         if haskey(meta, TRACKING_ID) && haskey(agent.transaction_handler, meta[TRACKING_ID])
             caller, response_handler = agent.transaction_handler[meta[TRACKING_ID]]
             delete!(agent.transaction_handler, meta[TRACKING_ID])
@@ -293,6 +332,29 @@ function get_model_handle(agent::Agent, type::DataType)
 end
 
 """
+Add a rule for message forwarding.
+
+After calling the agent will auto-forward every message coming from `from_addr` to
+`to_address`. If forward_replies is set, all replies from `to_address` are forwarded
+back to `from_addr`.
+"""
[…]
