+++ src/agent/api.jl
@@ -73,6 +73,19 @@ function send_tracked_message(
 end
 
 """
+API Definition for sending tracked message with required response handler
+"""
+function send_and_handle_answer(
+    response_handler::Function,
+    agent::AgentInterface,
+    content::Any,
+    agent_address::AgentAddress;
+    calling_object::Any=nothing,
+    kwargs...)
+    @warn "The API send_and_handle_answer definition has been called, this should never happen. There is most likely an import error."
+end
+
+"""
 API Definition for directly replying to a message
 """
 function reply_to(
@@ -81,4 +94,15 @@ function reply_to(
     received_meta::AbstractDict
 )
     @warn "The API reply_to definition has been called, this should never happen. There is most likely an import error."
+end
+
+"""
+API Definition for forwarding messages to other agents. 
+"""
+function forward_to(agent::AgentInterface,
+    content::Any,
+    forward_to_address::AgentAddress,
+    received_meta::AbstractDict;
+    kwargs...)    
+    @warn "The API forward_to definition has been called, this should never happen. There is most likely an import error."
 end

[…]
