### 2024-08-13 a50c572
Adding tests for rule deletion and send and handle.
