+++ src/agent/api.jl
@@ -1,4 +1,4 @@
-export Agent, AgentInterface, subscribe_message_handle, subscribe_send_handle, subscribe_event_handle, emit_event_handle, get_model_handle, send_message, send_tracked_message, reply_to, address, aid
+export Agent, AgentInterface, subscribe_message_handle, subscribe_send_handle, subscribe_event_handle, emit_event_handle, get_model_handle, send_message, send_tracked_message, reply_to, address, aid, send_and_handle_answer
 
 
 """
