+++ previews/PR48/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"role/#Roles","page":"Roles","title":"Roles","text":"","category":"section"},{"location":"role/","page":"Roles","title":"Roles","text":"Roles are used to provide a mechanism for reusability and modularization of functionsalities provided/implemented by agents. Every agent can contain and unlimited number of roles, which are separate structs on which typical agent functionalitites (like send_message) can be defined. All roles of an agent share the same address and agent id, as they are part of the agent and no autonomous unit for themself. ","category":"page"},{"location":"role/#Role-definition","page":"Roles","title":"Role definition","text":"","category":"section"},{"location":"role/","page":"Roles","title":"Roles","text":"A role can be defined using the @role macro. This macro adds some baselinefields to the following struct definition. The struct can be defined like any other Julia struct.","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"# Define your role struct using @role macro\n@role struct MyRole\n    my_own_field::String\nend\n\n# Assume you have already defined roles using Mango.AgentRole module\nrole1 = MyRole(\"Role1\")","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"Most functions, used for agent development can also be used with roles (e.g. handle_message, address, schedule and the lifecycle methods).  ","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"Additionally, roles can define the setup function to define actions to take when the roles are added to the agent. It is also possible to subscribe to specific messages using a boolean expression with the subscribe_message(role::Role, handler::Function, condition::Function) function. With the @role macro, the role context is added to the role, which contains the reference to the agent. However, it is recommended to use the equivalent methods defined on the role to execute actions like scheduling and sending messages. Further with roles it is possible to listen to all messages sent from within the agent. For this subscribe_send(role, handler::Function) can be used.","category":"page"},{"location":"role/#Role-communication","page":"Roles","title":"Role communication","text":"","category":"section"},{"location":"role/","page":"Roles","title":"Roles","text":"Besides the message subscriptions there are functionalities to communicate/work together with other roles. There are two different mechanisms for this:","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"Data sharing\nAn event system","category":"page"},{"location":"role/#Data-sharing","page":"Roles","title":"Data sharing","text":"","category":"section"},{"location":"role/","page":"Roles","title":"Roles","text":"The data sharing can be used using ordinary Julia structs with default constructors. There are two ways to share the data, first you can create the model you want share with get_model(role, your_data_model_type)","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"struct TestModel\n    c::Int64\nend\nTestModel() = TestModel(42)\nrole = MyRole(\"Role1\")\nshared_model = get_model(role, TestModel)","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"Mango.jl will create a TestModel instance and manage this instance such that every role can access it. ","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"Although this is a straightforward method it can be very clumsy to use. For this reason there is the macro @shared, which can be used within a role definition to mark a field as shared model. Then, Mango.jl will ensure that a shared instance of the declared type will be created and assigned to the struct field.","category":"page"},{"location":"role/","page":"Roles","title":"Roles","text":"@role struct SharedFieldTestRole\n    @shared \n    test_model::TestModel\nend","category":"page"},{"location":"role/#Event-system","page":"Roles","title":"Event system","tex
[…]
