### 2024-08-13 #7098 @aurorarossi
Hello,

I’ve started reviewing the package, and I’ve noticed the following:

- [ ]  **Missing API List**: The documentation lacks a clear list of available APIs.

- [ ]   **Non-standard Docstrings**: The docstrings within the package do not adhere to the standard Julia conventions. I recommend adapting them to the [Julia documentation standards](https://docs.julialang.org/en/v1/manual/documentation/) for better consistency and easier integration into the overall documentation, especially in the API section. For example, the documentation describes only two fields of the `Container` structure, but there are actually seven fields in the code, none of which are explained in detail.

- [ ]  **Example Code Issues**: The example provided in the `README.md`, which also appears in the paper and documentation, fails to work when copied directly into my terminal. Specifically, it is missing `using Sockets` and `import Sockets: InetAddr`. Additionally, there’s an issue with the `send_message` function (see the attached screenshot of the error ).
![image](https://github.com/user-attachments/assets/429470dc-c059-4247-9e30-3ca49ec3f578)
