### 2024-08-13 8a10266
Adding readme to CI and fixing related bug.
