+++ src/agent/core.jl
@@ -146,6 +146,7 @@ agent.
 function dispatch_message(agent::Agent, message::Any, meta::AbstractDict)
     lock(agent.lock) do
         if haskey(meta, TRACKING_ID) && haskey(agent.transaction_handler, meta[TRACKING_ID])
+            @info "asd"
             caller, response_handler = agent.transaction_handler[meta[TRACKING_ID]]
             delete!(agent.transaction_handler, meta[TRACKING_ID])
             response_handler(caller, message, meta)
@@ -396,7 +397,7 @@ function send_tracked_message(
     agent::Agent,
     content::Any,
     agent_address::AgentAddress;
-    response_handler::Function=(agent,message,meta)->nothing,
+    response_handler::Union{Function,Nothing}=nothing,
     calling_object::Any=nothing,
     kwargs...
 )
@@ -427,13 +428,12 @@ container.
 function reply_to(agent::Agent,
     content::Any,
     received_meta::AbstractDict;
-    response_handler::Function=(agent,message,meta)->nothing,
+    response_handler::Union{Function,Nothing}=nothing,
     calling_object::Any=nothing,
     kwargs...)
-    target_meta = Dict(received_meta)
-    return send_tracked_message(agent, content, AgentAddress(target_meta[SENDER_ID], 
-                                              target_meta[SENDER_ADDR], 
-                                              target_meta[TRACKING_ID]); 
+    return send_tracked_message(agent, content, AgentAddress(received_meta[SENDER_ID], 
+                                              received_meta[SENDER_ADDR], 
+                                              get(received_meta, TRACKING_ID, nothing)); 
                                               response_handler=response_handler,
                                               calling_object=calling_object,
                                               kwargs...)
