### 2024-08-13 af9a522
make example copy-pastable from start to finish
