### 2024-08-13 #7098 @jsagerOffis
@aurorarossi Hi, thanks for the inputs.

API List and Docstring standards: we will get on this and then merge it back over from the development branch.

Examples:
* README.md also had a bug that is now fixed an development and will be merged over later.
* The example in the paper is now updated so it can be copy pasted from start to finish and should produce a working program.

Regarding the `send_message` bug:
Going by the second suggested method candidate, it looks like you have a local definition of the `AgentAddress` struct so it can not resolve the correct method.
I assume this happened by copying the struct info we put in the corresponding part of the documentation?

Otherwise the `send_message` call looks like it should function.
