### 2024-08-14 #7098 @aurorarossi
@jsagerOffis I tried again, copying and pasting the example from the paper. The `send_message` function works now, but it seems to be stuck in a loop.
![image](https://github.com/user-attachments/assets/7d38f757-1a40-400a-be70-12fdd01d3d1f)

Could the issue be that both containers have the same name? Shouldn't each container/agent have a unique name?
![image](https://github.com/user-attachments/assets/58862d7e-7af0-4d2d-bedd-f9d00c9d9ee8)
