### 2024-08-14 cc3076c
change paper example to have nicer outputs
