### 2024-08-14 #7098 @jsagerOffis
I added a delay to the example agents answering so they spam less and also changed their names and outputs to make things a little clearer.

![example_output](https://github.com/user-attachments/assets/2c1d6044-2512-4ffe-a55c-c50c67b23a55)
