### 2024-08-14 #7098 @aurorarossi
@jsagerOffis, I've identified the issue. When I copy and paste the following block from your paper:
```julia
# Start the container
wait(Threads.@spawn start(container))
wait(Threads.@spawn start(container2))

# Send the first message to start the exchange
send_message(ping_agent, "Ping", address(pong_agent))
```
it runs in a loop because it lacks the second part, which is in a different markdown block:
```julia
# wait a bit to see some outputs
sleep(5)

@sync begin
    Threads.@spawn shutdown(container)
    Threads.@spawn shutdown(container2)
end
```
I strongly suggest creating a function that encapsulates both blocks of code.
