### 2024-08-14 4a1ce8f
Adding run_mango convenience function.
