+++ test/runtests.jl
@@ -8,7 +8,7 @@ using Test
     include("container_tests.jl")
     include("encode_decode_tests.jl")
     include("simulation_container_tests.jl")
-    include("examples.jl")
+    #include("examples.jl")
     include("tcp_protocol_tests.jl")
     include("agent_modeling_tests.jl")
 end
