+++ src/agent/core.jl
@@ -146,7 +146,6 @@ agent.
 function dispatch_message(agent::Agent, message::Any, meta::AbstractDict)
     lock(agent.lock) do
         if haskey(meta, TRACKING_ID) && haskey(agent.transaction_handler, meta[TRACKING_ID])
-            @info "asd"
             caller, response_handler = agent.transaction_handler[meta[TRACKING_ID]]
             delete!(agent.transaction_handler, meta[TRACKING_ID])
             response_handler(caller, message, meta)
@@ -431,9 +430,10 @@ function reply_to(agent::Agent,
     response_handler::Union{Function,Nothing}=nothing,
     calling_object::Any=nothing,
     kwargs...)
-    return send_tracked_message(agent, content, AgentAddress(received_meta[SENDER_ID], 
-                                              received_meta[SENDER_ADDR], 
-                                              get(received_meta, TRACKING_ID, nothing)); 
+    target_meta = Dict(received_meta)
+    return send_tracked_message(agent, content, AgentAddress(target_meta[SENDER_ID], 
+                                              target_meta[SENDER_ADDR], 
+                                              get(target_meta, TRACKING_ID, nothing)); 
                                               response_handler=response_handler,
                                               calling_object=calling_object,
                                               kwargs...)
