### 2024-08-14 24b45de
Adding exception handling for run_mango.
