### 2024-08-14 57b340b
Exclude MQTT if broker not connected (!= not linux test environment).
