### 2024-08-15 #7098 @jsagerOffis
> I strongly suggest creating a function that encapsulates both blocks of code.

Great suggestion. @rcschrg is currently adding this with syntax like:
```Julia
activate(containers) do
    # Send the first message to start the exchange
    send_message(ping_agent, "Ping", address(pong_agent))

    # wait a bit to see some outputs
    sleep(5)
end
```
