### 2024-08-16 ab69f33
Adding more convenience functions, documented, and tests for it.
