+++ src/container/simulation.jl
@@ -12,228 +12,228 @@ AGENT_PREFIX::String = "agent"
 # DISCRETE EVENT STEP SIZE
 DISCRETE_EVENT::Real = -1
 
-function create_simulation_container(start_time::DateTime; communication_sim::Union{Nothing,CommunicationSimulation}=nothing, task_sim::Union{Nothing,TaskSimulation}=nothing) 
-    container = SimulationContainer()
-    container.clock.simulation_time = start_time
-    if !isnothing(communication_sim)
-        container.communication_sim = communication_sim
-    end
-    if !isnothing(task_sim)
-        container.task_sim = task_sim
-    end
-    return container
+function create_simulation_container(start_time::DateTime; communication_sim::Union{Nothing, CommunicationSimulation} = nothing, task_sim::Union{Nothing, TaskSimulation} = nothing)
+	container = SimulationContainer()
+	container.clock.simulation_time = start_time
+	if !isnothing(communication_sim)
+		container.communication_sim = communication_sim
+	end
+	if !isnothing(task_sim)
+		container.task_sim = task_sim
+	end
+	return container
 end
 
 struct MessageData
-    content::Any
-    meta::AbstractDict
-    arriving_time::DateTime
+	content::Any
+	meta::AbstractDict
+	arriving_time::DateTime
 end
 
 @kwdef mutable struct SimulationContainer <: ContainerInterface
-    world::World=World()
-    clock::Clock=Clock(DateTime(0))
-    task_sim::TaskSimulation=SimpleTaskSimulation(clock=clock)
-    agents::Dict{String,Agent} = Dict()
-    agent_counter::Integer = 0
-    shutdown::Bool = false
-    communication_sim::CommunicationSimulation = SimpleCommunicationSimulation()
-    message_queue::ConcurrentQueue{MessageData} = ConcurrentQueue{MessageData}()
+	world::World = World()
+	clock::Clock = Clock(DateTime(0))
+	task_sim::TaskSimulation = SimpleTaskSimulation(clock = clock)
+	agents::Dict{String, Agent} = Dict()
+	agent_counter::Integer = 0
+	shutdown::Bool = false
+	communication_sim::CommunicationSimulation = SimpleCommunicationSimulation()
+	message_queue::ConcurrentQueue{MessageData} = ConcurrentQueue{MessageData}()
 end
 
 function on_step(agent::Agent, world::World, clock::Clock, step_size_s::Real)
-    # default nothing
+	# default nothing
 end
 
 function on_step(role::Role, world::World, clock::Clock, step_size_s::Real)
-    # default nothing
+	# default nothing
 end
 
 function step_agent(agent::Agent, world::World, clock::Clock, step_size_s::Real)
-    on_step(agent, world, clock, step_size_s)
-    for role in roles(agent)
-        on_step(role, world, clock, step_size_s)
-    end
+	on_step(agent, world, clock, step_size_s)
+	for role in roles(agent)
+		on_step(role, world, clock, step_size_s)
+	end
 end
 
 struct MessagingIterationResult
-    communication_result::CommunicationSimulationResult
-    state_changed::Bool
+	communication_result::CommunicationSimulationResult
+	state_changed::Bool
 end
 @kwdef struct MessagingSimulationResult
-    results::Vector{MessagingIterationResult} = Vector()
+	results::Vector{MessagingIterationResult} = Vector()
 end
 @kwdef struct TaskSimulationResult
-    results::Vector{TaskIterationResult} = Vector()
+	results::Vector{TaskIterationResult} = Vector()
 end
 
 struct SimulationResult
-    time_elapsed::Real
-    messasing_result::MessagingSimulationResult
-    task_result::TaskSimulationResult
[…]
