+++ src/util/scheduling.jl
@@ -1,18 +1,18 @@
 export TaskData,
-    PeriodicTaskData,
-    InstantTaskData,
-    DateTimeTaskData,
-    AwaitableTaskData,
-    ConditionalTaskData,
-    execute_task,
-    stop_task,
-    stop_all_tasks,
-    wait_for_all_tasks,
-    stop_and_wait_for_all_tasks,
-    schedule,
-    Scheduler,
-    AbstractScheduler,
-    Clock
+	PeriodicTaskData,
+	InstantTaskData,
+	DateTimeTaskData,
+	AwaitableTaskData,
+	ConditionalTaskData,
+	stop_task,
+	stop_all_tasks,
+	wait_for_all_tasks,
+	stop_and_wait_for_all_tasks,
+	schedule,
+	Clock,
+	AbstractScheduler,
+	Scheduler,
+	SimulationScheduler
 
 using Dates
 using ConcurrentCollections
@@ -22,7 +22,7 @@ import Base.schedule, Base.sleep, Base.wait
 abstract type AbstractClock end
 
 @kwdef mutable struct Clock <: AbstractClock
-    simulation_time::DateTime
+	simulation_time::DateTime
 end
 
 struct DateTimeClock <: AbstractClock
@@ -35,146 +35,146 @@ abstract type TaskData end
 abstract type AbstractScheduler end
 
 function now(scheduler::AbstractScheduler)
-    return DateTime.now()
+	return DateTime.now()
 end
 
 function sleep(scheduler::AbstractScheduler, time_s::Real)
-    return sleep(time_s)
+	return sleep(time_s)
 end
 
 function wait(scheduler::AbstractScheduler, timer::Timer, delay_s::Real)
-    return wait(timer)
+	return wait(timer)
 end
 
 function clock(scheduler::AbstractScheduler)
-    throw("unimplemented")
+	throw("unimplemented")
 end
 function tasks(scheduler::AbstractScheduler)
-    throw("unimplemented")
+	throw("unimplemented")
 end
-    
+
 @kwdef struct Scheduler <: AbstractScheduler
-    tasks::AbstractDict{Task,TaskData} = ConcurrentDict{Task,TaskData}()
-    clock::DateTimeClock = DateTimeClock()
+	tasks::AbstractDict{Task, TaskData} = ConcurrentDict{Task, TaskData}()
+	clock::DateTimeClock = DateTimeClock()
 end
 
 function clock(scheduler::Scheduler)
-    return scheduler.clock
+	return scheduler.clock
 end
 function tasks(scheduler::Scheduler)
-    return scheduler.tasks
+	return scheduler.tasks
 end
-        
+
 function is_stopable(data::TaskData)::Bool
-    return false
+	return false
 end
 
 function stop_single_task(data::TaskData)::Nothing
 end
 
 mutable struct PeriodicTaskData <: TaskData
-    interval_s::Real
-    condition::Function
-    timer::Timer
+	interval_s::Real
+	condition::Function
+	timer::Timer
 
-    function PeriodicTaskData(interval_s::Real, condition::Function=()->true)
-        return new(interval_s, condition, Timer(interval_s; interval=interval_s))
-    end
+	function PeriodicTaskData(interval_s::Real, condition::Function = () -> true)
+		return new(interval_s, condition, Timer(interval_s; interval = interval_s))
+	end
 end
 
 function is_stopable(data::PeriodicTaskData)::Bool
-    return true
+	return true
 end
 
 function stop_single_task(data::PeriodicTaskData)::Nothing
-    close(data.timer)
+	close(data.timer)
 end
 
 struct InstantTaskData <: TaskData end
 
[…]
