+++ src/container/protocol.jl
@@ -1,5 +1,5 @@
 
-export Protocol, send, start, close, id, notify_register
+export Protocol, send, start, close, id, notify_register, init
 
 """
 Type for all implementations of protocols, acts like an interface. A protocol defines the way message are processed and especially sent and received 
@@ -11,8 +11,8 @@ The parameterized type T indicates type, which defines the address data of the r
 Every protocol has to define two methods.
 1. send: defines the behavior of the protocol when an agents sends a messages
 2. init: defines the necessary steps to initialize (especially) the receiver loop and
-         therefore accepts a stop check and a data handler function to indicate when the receiver should stop, 
-         respectively how to dispatch the received message to the correct agent
+		 therefore accepts a stop check and a data handler function to indicate when the receiver should stop, 
+		 respectively how to dispatch the received message to the correct agent
 """
 abstract type Protocol{T} end
 
