+++ src/agent/api.jl
@@ -1,5 +1,10 @@
-export Agent, AgentInterface, subscribe_message_handle, subscribe_send_handle, subscribe_event_handle, emit_event_handle, get_model_handle, send_message, send_tracked_message, reply_to, address, aid
+export Agent, send_message, send_tracked_message, reply_to, address, aid
 
+
+"""
+Supertype of all address types
+"""
+abstract type Address end
 
 """
 Supertype of the Agent Base type representing an interface, on which methods
@@ -55,7 +60,7 @@ API Definition for the role context
 function send_message(
     agent::AgentInterface,
     content::Any,
-    agent_address::AgentAddress;
+    agent_address::Address;
     kwargs...
 )
     @warn "The API send_message definition has been called, this should never happen. There is most likely an import error."
@@ -64,7 +69,7 @@ end
 function send_tracked_message(
     agent::AgentInterface,
     content::Any,
-    agent_address::AgentAddress;
+    agent_address::Address;
     response_handler::Function=(agent,message,meta)->nothing,
     calling_object::Any=nothing,
     kwargs...

[…]
