+++ test/examples.jl
@@ -99,7 +99,7 @@ end
 		end
 
 		# Send the first message to start the exchange
-		wait(send_message(ping_agent, "Ping", MQTTAddress(broker_addr, "pings")))
+		wait(send_message(ping_agent, "Ping", MQTTAddress(InetAddr("127.0.0.1", 1883), "pings")))
 
 		# Wait for a moment to see the result
 		# In general you want to use a Condition() instead to
