+++ src/agent/core.jl
@@ -1,18 +1,18 @@
 export @agent,
-	AgentContext,
-	AgentRoleHandler,
-	handle_message,
-	add,
-	schedule,
-	stop_and_wait_for_all_tasks,
-	shutdown,
-	on_ready,
-	on_start,
-	roles,
-	forward_to,
-	add_forwarding_rule,
-	delete_forwarding_rule,
-	ForwardingRule
+    AgentContext,
+    AgentRoleHandler,
+    handle_message,
+    add,
+    schedule,
+    stop_and_wait_for_all_tasks,
+    shutdown,
+    on_ready,
+    on_start,
+    roles,
+    forward_to,
+    add_forwarding_rule,
+    delete_forwarding_rule,
+    ForwardingRule
 
 using UUIDs
 
@@ -27,24 +27,24 @@ connection to the container, including all functions used for interacting with t
 for the agent.
 """
 struct AgentContext
-	container::ContainerInterface
+    container::ContainerInterface
 end
 
 """
 Internal data regarding the roles.
 """
 struct AgentRoleHandler
-	roles::Vector{Role}
-	handle_message_subs::Vector{Tuple{Role, Function, Function}}
-	send_message_subs::Vector{Tuple{Role, Function}}
-	event_subs::Dict{Any, Vector{Tuple{Role, Function, Function}}}
-	models::Dict{DataType, Any}
+    roles::Vector{Role}
+    handle_message_subs::Vector{Tuple{Role,Function,Function}}
+    send_message_subs::Vector{Tuple{Role,Function}}
+    event_subs::Dict{Any,Vector{Tuple{Role,Function,Function}}}
+    models::Dict{DataType,Any}
 end
 
 struct ForwardingRule
-	from_address::AgentAddress
-	to_address::AgentAddress
-	forward_replies::Bool
+    from_address::AgentAddress
+    to_address::AgentAddress
+    forward_replies::Bool
 end
 
 """
@@ -52,13 +52,13 @@ All baseline fields added by the @agent macro are listed in this vector.
 They are added in the same order defined here.
 """
 AGENT_BASELINE_FIELDS::Vector = [
-	:(lock::ReentrantLock),
-	:(context::Union{Nothing, AgentContext}),
-	:(role_handler::Union{AgentRoleHandler}),
-	:(scheduler::AbstractScheduler),
-	:(aid::Union{Nothing, String}),
-	:(transaction_handler::Dict{String, Tuple}),
-	:(forwarding_rules::Vector{ForwardingRule}),
+    :(lock::ReentrantLock),
+    :(context::Union{Nothing,AgentContext}),
+    :(role_handler::Union{AgentRoleHandler}),
+    :(scheduler::AbstractScheduler),
+    :(aid::Union{Nothing,String}),
+    :(transaction_handler::Dict{String,Tuple}),
+    :(forwarding_rules::Vector{ForwardingRule}),
 ]
 
 """
@@ -67,13 +67,13 @@ an anonymous functions. Always need to have the same length as
 AGENT_BASELINE_FIELDS.
 """
 AGENT_BASELINE_DEFAULTS::Vector = [
-	() -> ReentrantLock(),
-	() -> nothing,
-	() -> AgentRoleHandler(Vector(), Vector(), Vector(), Dict(), Dict()),
-	() -> Scheduler(),
-	() -> nothing,
-	() -> Dict{String, Tuple}(),
-	() -> Vector{ForwardingRule}(),
+    () -> ReentrantLock(),
+    () -> nothing,
+    () -> AgentRoleHandler(Vector(), Vector(), Vector(), Dict(), Dict()),
+    () -> Scheduler(),
+    () -> nothing,
[…]
