+++ src/util/error_handling.jl
@@ -1,12 +1,12 @@
 
 macro spawnlog(expr)
-	quote
-		Threads.@spawn try
-			$(esc(expr))
-		catch ex
-			bt = stacktrace(catch_backtrace())
-			showerror(stderr, ex, bt)
-			rethrow(ex)
-		end
-	end
+    quote
+        Threads.@spawn try
+            $(esc(expr))
+        catch ex
+            bt = stacktrace(catch_backtrace())
+            showerror(stderr, ex, bt)
+            rethrow(ex)
+        end
+    end
 end

[…]
