### 2024-08-17 #7098 @Tortar
I'm not sure that these are the same issues @aurorarossi found but I tried to run the example in the ReadMe but it errors at some point and then it runs in a loop (maybe)

```julia
julia> wait(send_message(ping_agent, "Ping", AgentAddress(aid=pong_agent.aid, address=InetAddr(ip"127.0.0.1", 2939))))

ERROR: UndefVarError: `AgentAddress` not defined in `Main`
Suggestion: check for spelling errors or missing imports.
Stacktrace:
 [1] top-level scope
   @ REPL[16]:1

# this never finishes?
julia> wait(Threads.@spawn begin
           while ping_agent.counter < 5
               sleep(1)
           end
       end)
```
