### 2024-08-17 eb24c3c
First version of topology.
