+++ Project.toml
@@ -2,16 +2,18 @@ name = "Mango"
 uuid = "5e49fdec-d473-4d14-b295-7bff2fcf1925"
 authors = ["OFFIS e.V."]
 repo = "https://github.com/OFFIS-DAI/Mango.jl"
-version = "0.3"
+version = "0.3.0"
 
 [deps]
 ConcurrentCollections = "5060bff5-0b44-40c5-b522-fcd3ca5cecdd"
 ConcurrentUtilities = "f0e56b4a-5159-44fe-b623-3e5288b988bb"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
+Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 LightBSON = "a4a7f996-b3a6-4de6-b9db-2fa5f350df41"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
+MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
 Mosquitto = "db317de6-444b-4dfa-9d0e-fbf3d8dd78ea"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
@@ -23,10 +25,12 @@ ConcurrentCollections = "~0.1"
 ConcurrentUtilities = "~2.2"
 JSON = "~0.21"
 LightBSON = "~0.2"
+Mosquitto = "~0.10"
 OrderedCollections = "~1.6"
 Parameters = "~0.12"
+Graphs = "~1.11"
+MetaGraphsNext = "~0.7"
 julia = "^1.9"
-Mosquitto = "~0.10"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
