### 2024-08-19 #7098 @jsagerOffis
@Tortar Well, if the first part there errors, its unsurprising that the second part can not finish.

The example setup is:
* We have two agents set to automatically reply to all incoming messages and counting them.
* After the initial message, this means they will indefinitely message each other. There is no other stop condition.

```Julia
wait(Threads.@spawn begin
           while ping_agent.counter < 5
               sleep(1)
           end
     end)
```
This waits for the ping agent to have received 5 mesages but since the initial message errored, this never happens.
As for why the initial send message errors to begin with... I actually don't know.

All the examples will slightly change today once I reviewed the new feature @rcschrg pushed so I hope that will make the example simpler.

Until then I just made this small mash-up of the paper and README.md examples (which I will turn into the same example for consistency later).
This does the same thing as the README.md example but with nicer outputs and with the same stopping logic.
Just ran this and it worked as expected so please let me know if it causes an issue for you.

Note that if you paste this into the REPL piece by piece it will start printing messages indefinitely after
`send_message(ping_agent, "Ping", address(pong_agent))` but should still be responsive to the remaining commands to shut it down (because agent logic is threaded in the background).

```Julia
using Mango
using Sockets: InetAddr, @ip_str

# Create the container instances with TCP protocol
container = Container()
container.protocol = TCPProtocol(address=InetAddr(ip"127.0.0.1", 5555))

container2 = Container()
container2.protocol = TCPProtocol(address=InetAddr(ip"127.0.0.1", 5556))

@agent struct TCPPingPongAgent
    counter::Int
end

# Create instances of ping pong agents
ping_agent = TCPPingPongAgent(0)
pong_agent = TCPPingPongAgent(0)

# register each agent to a container and give them a name
register(container, ping_agent, "Agent_1")
register(container2, pong_agent, "Agent_2")

import Mango.handle_message

# Override the default handle_message function for ping pong agents
function handle_message(agent::TCPPingPongAgent, message::Any, meta::Any)
    println(
        "$(agent.aid) got a message: $message." *
        "This is message number: $(agent.counter) for me!"
    )

    # doing very important work
    sleep(0.5)

    if message == "Ping"
        agent.counter += 1
        t = AgentAddress(meta["sender_id"], meta["sender_addr"], nothing)
        send_message(agent, "Pong", t)
    elseif message == "Pong"
        agent.counter += 1
        t = AgentAddress(meta["sender_id"], meta["sender_addr"], nothing)
        send_message(agent, "Ping", t)
    end
end

# Start the container
wait(Threads.@spawn start(container))
wait(Threads.@spawn start(container2))

# Send the first message to start the exchange
send_message(ping_agent, "Ping", address(pong_agent))

while ping_agent.counter < 5
        sleep(1)
end

@sync begin
    Threads.@spawn shutdown(container)
    Threads.@spawn shutdown(container2)
end
```
