### 2024-08-19 #7098 @jsagerOffis
@Tortar Are you running the latest version of Mango.jl? `AgentAddress` on top level was only added in 0.3.0 iirc. We pushed that release prior to submitting the paper but maybe you got 0.2.x for some reason anyway?
