### 2024-08-19 e1a9f1d
Update getting_started.md
