+++ src/container/core.jl
@@ -21,7 +21,7 @@ define the behavior of the container itself. That being said, the same container
 able to send messages via different protocols using different codecs.
 """
 @kwdef mutable struct Container <: ContainerInterface
-    agents::Dict{String,Agent} = Dict()
+    agents::OrderedDict{String,Agent} = OrderedDict()
     agent_counter::Integer = 0
     protocol::Union{Nothing,Protocol} = nothing
     codec::Any = (encode, decode)
@@ -30,8 +30,8 @@ able to send messages via different protocols using different codecs.
     tasks::Any = nothing
 end
 
-function agents(container::Container)
-    return values(container.agents)
+function agents(container::Container)::Vector{Agent}
+    return [t[2] for t in collect(container.agents)]
 end
 
 """
