+++ src/world/topology.jl
@@ -1,7 +1,8 @@
-export complete_topology, star_topology, cycle_topology, graph_topology, per_node, add!, topology_neighbors
+export complete_topology, star_topology, cycle_topology, graph_topology, per_node, add!, topology_neighbors, create_topology, add_node!, add_edge!
 
 using MetaGraphsNext
 using Graphs
+import Graphs.add_edge!
 
 @kwdef struct Node
     id::Int
@@ -20,8 +21,8 @@ function neighbors(service::TopologyService)
     return service.neighbors
 end
 
-function _create_meta_graph_with(graph::Graph)
-    vertices_description = [i => Node(i) for i in vertices(graph)]
+function _create_meta_graph_with(graph::AbstractGraph)
+    vertices_description = [i => Node(id=i) for i in vertices(graph)]
     edges_description = [
         (e.src, e.dst) => nothing for e in edges(graph)
     ]
@@ -30,37 +31,54 @@ function _create_meta_graph_with(graph::Graph)
 end
 
 function complete_topology(number_of_nodes::Int)
-    complete_graph = complete_graph(number_of_nodes)
-    return Topology(_create_meta_graph_with(complete_graph))
+    graph = complete_graph(number_of_nodes)
+    return Topology(_create_meta_graph_with(graph))
 end
 
 function star_topology(number_of_nodes::Int)
-    complete_graph = star_graph(number_of_nodes)
-    return Topology(_create_meta_graph_with(complete_graph))
+    graph = star_graph(number_of_nodes)
+    return Topology(_create_meta_graph_with(graph))
 end
 
 function cycle_topology(number_of_nodes::Int)
-    complete_graph = cycle_graph(number_of_nodes)
-    return Topology(_create_meta_graph_with(complete_graph))
+    graph = cycle_graph(number_of_nodes)
+    return Topology(_create_meta_graph_with(graph))
 end
 
 function graph_topology(graph::Graph)
     return Topology(_create_meta_graph_with(graph))
 end
 
-function per_node(assign_runnable::Function, topology::Topology)
-    # 1st pass, let the user assign the agents
-    for label in labels(topology.graph)
-        node = topology.graph[label]
-        assign_runnable(node)
+function add_edge!(topology::Topology, node_id_from::Int, node_id_to::Int, directed=false)
+    if directed
+        topology.graph[node_id_from, node_id_to] = nothing
+    else
+        topology.graph[node_id_to, node_id_from] = nothing
+        topology.graph[node_id_from, node_id_to] = nothing
     end
+end
+
+function add_node!(topology::Topology, agents::Agent...)::Int
+    vid = nv(topology.graph) + 1
+    topology.graph[vid] = Node(vid, [a for a in agents])
+    return vid
+end
+
+function create_topology(create_runnable::Function)
+    topology = Topology(_create_meta_graph_with(DiGraph()))
+    create_runnable(topology)
+    _build_neighborhoods_and_inject(topology)
+    return topology
+end
+
+function _build_neighborhoods_and_inject(topology::Topology)
     # 2nd pass, build the neighborhoods and add it to agents
     for label in labels(topology.graph)
         node = topology.graph[label]
-        neighbor_addresses = Vector()
+        neighbor_addresses::Vector{AgentAddress} = Vector{AgentAddress}()
         for n_label in neighbor_labels(topology.graph, label)
             n_node = topology.graph[n_label]
-            push!(neighbor_addresses, [address(node) for agent in n_node.agents])
+            append!(neighbor_addresses, [address(agent) for agent in n_node.agents])
         end
         for agent in node.agents
             topology_service = service_of_type(agent, TopologyService, TopologyService())
[…]
