### 2024-08-19 f08b710
Adding docstrings.
