+++ previews/PR57/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"scheduling/#Scheduling","page":"Scheduling","title":"Scheduling","text":"","category":"section"},{"location":"scheduling/#.-Introduction","page":"Scheduling","title":"1. Introduction","text":"","category":"section"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"Welcome to the documentation for the Scheduling component in Mango.jl. This utility component provides a flexible scheduler for executing predefined tasks. It offers various TaskData types to specify different task execution behaviors.","category":"page"},{"location":"scheduling/#.-Module-Overview","page":"Scheduling","title":"2. Module Overview","text":"","category":"section"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"The Scheduling module exports several types and functions to facilitate task scheduling and execution. Let's briefly review the main components of this module.","category":"page"},{"location":"scheduling/#Task-Data-Types","page":"Scheduling","title":"Task Data Types","text":"","category":"section"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"The module provides different TaskData types, each catering to specific scheduling requirements:","category":"page"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"PeriodicTaskData: For tasks that need to be executed periodically, it holds the time interval in seconds between task executions.\nInstantTaskData: For tasks that need to be executed instantly, without any delay.\nDateTimeTaskData: For tasks that need to be executed at a specific date and time.\nAwaitableTaskData: For tasks that require waiting for an awaitable object to complete before execution.\nConditionalTaskData: For tasks that execute based on a specific condition at regular intervals.","category":"page"},{"location":"scheduling/#Typical-usage","page":"Scheduling","title":"Typical usage","text":"","category":"section"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"Typically the scheduler is used within methods from the agent. To schedule a task the function schedule can be used. It takes two inputs: The agent (which forwards the call to its scheduler) and the TaskData object of the task.","category":"page"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"agent = MyAgent(0)\nresult = 0\n\nschedule(agent, InstantTaskData()) do \n    # some expensive calculation\n    result = 10       \nend\nwait_for_all_tasks(agent)","category":"page"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"PeriodicTaskData creates tasks that get executed repeatedly forever.  This means that calling wait on such a task will generally simply block forever. For this reason a periodic task has to be stopped before it can be waited on.","category":"page"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"delay_in_s = 0.5 # delay between executions of the task in seconds\n\nt = schedule(agent, PeriodicTaskData(delay)) do \n    # some expensive calculation\n    result = 10       \nend\n\nstop_task(agent, t)\nwait_for_all_tasks(agent)","category":"page"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"Alternatively, you can stop all stopable tasks simultaneously with the stop_all_tasks function.","category":"page"},{"location":"scheduling/","page":"Scheduling","title":"Scheduling","text":"delay_in_s = 0.5 # delay between executions of the task in seconds\n\nfor i in 1:100\n    schedule(agent, PeriodicTaskData(delay)) do \n        # some expensive calculation\n        result = 10       \n    end\nend\n\nstop_all_task(agent, t)\nwait_for_all_tasks(agent)","category":"page"},{"location":"scheduling/","page":"Scheduling","title"
[…]
