### 2024-08-20 b8922fe
Adding missing tests for coverage.
