### 2024-08-20 d8f40a7
Typos topology.md
