### 2024-08-20 06d1ad6
Added autodocs as test.
