### 2024-08-20 fd9ece5
add codec docstrings and split encode type union
