### 2024-08-20 45628f6
remove unused old struct
