### 2024-08-20 d10cac0
add docstrings to mqtt.jl
