### 2024-08-20 bb31355
Docstrings for agent core and api.
