+++ src/agent/core.jl
@@ -93,6 +93,7 @@ The macro does 3 things:
    to predefined default values. As a result you can (and should) create 
    an agent using only the exclusive fields.
 
+# Example
 For example the usage could like this.
 ```julia
 @agent struct MyAgent
@@ -208,6 +209,8 @@ function dispatch_message(agent::Agent, message::Any, meta::AbstractDict)
 end
 
 """
+    handle_message(agent::Agent, message::Any, meta::Any)
+
 Defines a function for an agent, which will be called when a message is dispatched
 to the agent. This methods will be called with any arriving message (according to
 the multiple dispatch of julia).
@@ -231,6 +234,8 @@ function notify_ready(agent::Agent)
 end
 
 """
+    on_start(agent::Agent)
+
 Lifecycle Hook-in function called when the container of the agent has been started,
 depending on the container type it may not be called (if there is no start at all, 
 f.e. the simulation container)
@@ -240,6 +245,8 @@ function on_start(agent::Agent)
 end
 
 """
+    on_ready(agent::Agent)
+
 Lifecycle Hook-in function called when the agent system as a whole is ready, the 
 hook-in has to be manually activated using notify_ready(container::Container)
 """
@@ -247,14 +254,13 @@ function on_ready(agent::Agent)
     # do nothing by default
 end
 
-"""
-Returns the agent id of the agent.
-"""
 function aid(agent::Agent)
     return agent.aid
 end
 
 """
+    add(agent::Agent, role::Role)
+
 Add a role to the agent. This will add the role
 to the internal RoleHandler of the agent and it
 will bind the RoleContext to the role, which enables
@@ -266,6 +272,8 @@ function add(agent::Agent, role::Role)
 end
 
 """
+    roles(agent)
+
 Return all roles of the given agent
 """
 function roles(agent::Agent)
@@ -273,6 +281,8 @@ function roles(agent::Agent)
 end
 
 """
+    shutdown(agent)
+
 Will be called on shutdown of the container, in which
 the agent is living
 """
@@ -284,9 +294,6 @@ function shutdown(agent::Agent)
     stop_and_wait_for_all_tasks(agent.scheduler)
 end
 
-"""
-Internal implementation of the agent API.
-"""
 function subscribe_message_handle(
     agent::Agent,
     role::Role,
@@ -296,16 +303,10 @@ function subscribe_message_handle(
     push!(agent.role_handler.handle_message_subs, (role, condition, handler))
 end
 
-"""
-Internal implementation of the agent API.
-"""
 function subscribe_send_handle(agent::Agent, role::Role, handler::Function)
     push!(agent.role_handler.send_message_subs, (role, handler))
 end
 
-"""
-Internal implementation of the agent API.
-"""
 function subscribe_event_handle(agent::Agent, role::Role, event_type::Any, event_handler::Function; condition::Function=(a, b) -> true)
     if !haskey(agent.role_handler.event_subs, event_type)
         agent.role_handler.event_subs[event_type] = Vector()
@@ -313,9 +314,6 @@ function subscribe_event_handle(agent::Agent, role::Role, event_type::Any, event
     push!(agent.role_handler.event_subs[event_type], (role, condition, event_handler))
 end
 
-"""
-Internal implementation of the agent API.
-"""
 function emit_event_handle(agent::Agent, src::Role, event::Any; event_type::Any=nothing)
     key = !isnothing(event_type) ? event_type : typeof(event)
     if haskey(agent.role_handler.event_subs, key)
@@ -330,9 +328,6 @@ function emit_event_handle(agent::Agent, src::Role, event::Any; event_type::Any=
     end
 end
 
-"""
-Internal implementation of the agent API.
-"""
[…]
