+++ src/agent/api.jl
@@ -19,33 +19,40 @@ implementations across all agents.
 """
 abstract type Agent <: AgentInterface end
 
-"""
-Used internally by the RoleContext to subscribe message handler to the agent.
-"""
 function subscribe_message_handle(agent::AgentInterface, role::Any, condition::Any, handler::Any) end
 
 """
+    subscribe_send_handle(agent::AgentInterface, role::Any, handler::Any)
+
 Used internally by the RoleContext to subscribe send handler to the agent.
 """
 function subscribe_send_handle(agent::AgentInterface, role::Any, handler::Any) end
 
 """
+    subscribe_event_handle(agent::AgentInterface, role::Any, event_type::Any, event_handler::Any; condition::Function=(a, b) -> true)
+
 Used internally by the RoleContext to subscribe to role agent events.
 """
 function subscribe_event_handle(agent::AgentInterface, role::Any, event_type::Any, event_handler::Any; condition::Function=(a, b) -> true) end
 
 """
+    emit_event_handle(agent::AgentInterface, role::Any, event::Any; event_type::Any=nothing)
+
 Used internally by the RoleContext to subscribe to role agent events.
 """
 function emit_event_handle(agent::AgentInterface, role::Any, event::Any; event_type::Any=nothing) end
 
 """
+    get_model_handle(agent::AgentInterface, type::DataType)
+
 Used internally by the RoleContext to subscribe to role agent events.
 """
 function get_model_handle(agent::AgentInterface, type::DataType) end
 
 """
-Used internally by the role to get the AgentAddress
+    address(agent)
+
+Return the agent address of the agent as [`AgentAddress`](@ref) or [`MQTTAddress`](@ref) depending on the protocol.
 """
 function address(agent::AgentInterface) end
 
@@ -55,7 +62,12 @@ Used internally by the role to get the AID
 function aid(agent::AgentInterface) end
 
 """
-API Definition for the role context
+    send_message(agent, content::Any, agent_address::Address; kwargs...,)
+
+Send a message with the content `content` to the agent represented by `agent_address`. 
+
+This method will always set a sender_id. Additionally, further keyword arguments can be defined to fill the 
+internal meta data of the message.
 """
 function send_message(
     agent::AgentInterface,
@@ -66,6 +78,24 @@ function send_message(
     @warn "The API send_message definition has been called, this should never happen. There is most likely an import error."
 end
 
+
+"""
+    send_tracked_message(agent, content, agent_address; 
+        response_handler::Function(agent, message, meta)::nothing,
+        calling_object::Any=nothing,
+        kwargs...,
+    )
+
+Send a message with the content `content` to the agent represented by `agent_address`. This function will set
+a generated tracking_id to the address, which allows the identification of the dialog. 
+
+It is possible to define a `response_handler`, to which a function can be assigned, which handles the answer 
+to this message call. Note that the resonding agent needs to use the same tracking id in the response, ideally
+[`reply_to`](@ref) is used to achieve this automatically. 
+
+This method will always set a sender_id. Additionally, further keyword arguments can be defines to fill the 
+internal meta data of the message.
+"""
 function send_tracked_message(
     agent::AgentInterface,
     content::Any,
@@ -78,7 +108,22 @@ function send_tracked_message(
 end
 
 """
-API Definition for sending tracked message with required response handler
+    send_and_handle_answer(
+        response_handler::Function(agent, message, meta)::nothing,
+        agent::AgentInterface,
+        content::Any,
+        agent_address::Address;
+        calling_object::Any=nothing,
+        kwargs...)
+
+Convenience method for sending tracked messages with response handler to the answer.
+
+Sends a tracked message with a required response_handler to enable to use the syntax
+```
[…]
