### 2024-08-20 836e21c
add MQTTProtocol struct docstring
