### 2024-08-20 a90ad31
Documenting role apis.
