+++ src/agent/core.jl
@@ -248,7 +248,8 @@ end
     on_ready(agent::Agent)
 
 Lifecycle Hook-in function called when the agent system as a whole is ready, the 
-hook-in has to be manually activated using notify_ready(container::Container)
+hook-in has to be manually activated using notify_ready(container::Container). If you use
+    the `activate` function, it is called automatically.
 """
 function on_ready(agent::Agent)
     # do nothing by default
