### 2024-08-20 c520d98
Documenting most missing APIs.
