+++ src/container/core.jl
@@ -1,4 +1,4 @@
-export Container, register, send_message, start, shutdown, notify_ready
+export Container, notify_ready
 
 using Parameters
 using OrderedCollections
@@ -56,9 +56,6 @@ function process_message(container::Container, msg_data::Any, sender_addr::Any; 
     forward_message(container, content, meta; receivers=receivers)
 end
 
-"""
-Get protocol addr part
-"""
 function protocol_addr(container::Container)
     if isnothing(container.protocol)
         return nothing
@@ -66,10 +63,6 @@ function protocol_addr(container::Container)
     return id(container.protocol)
 end
 
-"""
-Starts the container and initialized all its components. After the call the container
-start to act as the communication layer.
-"""
 function start(container::Container)
     if !isnothing(container.protocol)
         container.loop, container.tasks = init(
@@ -84,7 +77,9 @@ function start(container::Container)
 end
 
 """
-Mark the agent system as ready, needs to be detected and called manually!
+    notify_ready(container::Container)
+
+Mark the agent system as ready.
 """
 function notify_ready(container::Container)
     for agent in values(container.agents)
@@ -92,9 +87,6 @@ function notify_ready(container::Container)
     end
 end
 
-"""
-Shut down the container. It is always necessary to call it for freeing bound resources
-"""
 function shutdown(container::Container)
     container.shutdown = true
     if !isnothing(container.protocol)
@@ -111,22 +103,6 @@ function shutdown(container::Container)
     end
 end
 
-"""
-Register an agent given the target container `container`. While registering
-an aid will be generated and assigned to the agent.
-
-This function will add the agent to the internal list of the container and will from
-then on be controlled by the container regarding the messaging activities. That means
-the container acts as the gateway of the agent defining its possible way to retrieve 
-messages.
-
-# Args
-suggested_aid: you can provide an aid yourself. The container will always use that aid
-	if possible
-
-# Returns
-The actually used aid will be returned.
-"""
 function register(
     container::Container,
     agent::Agent,
@@ -190,16 +166,6 @@ function to_external_message(content::Any, meta::AbstractDict)
     return MangoMessage(content, meta)
 end
 
-"""
-Send a message `message` with using the given container `container`
-to the agent with the receiver id `receiver_id`. The receivers address 
-is used by the chosen protocol to appropriatley route the message to
-external participants. To specifiy further meta data of the message
-`kwargs` should be used.
-
-# Returns
-True if the message has been sent successfully, false otherwise.
-"""
 function send_message(
     container::Container,
     content::Any,
@@ -240,6 +206,13 @@ end
 
 
 """
+    send_message(
+    container::Container,
+    content::Any,
+    mqtt_address::MQTTAddress,
+    kwargs...,
+)
+
 Send message version for MQTT topics. 
 Note that there is no local message forwarding here because messages always get
 pushed to a broker and are not directly addressed to an agennt.

[…]
