+++ src/container/api.jl
@@ -48,7 +48,7 @@ end
     kwargs...,
 )
 
-Send a message `message using the given container `container`
+Send a message `message` using the given container `container`
 to the given address. Additionally, further keyword
 arguments can be defines to fill the internal meta data of the message.
 """
