### 2024-08-21 5e50515
update README.md and paper.md with new QOL features in the example
