### 2024-08-21 fa714de
fix indentation in example wait loop
