+++ docs/src/agent.md
@@ -3,9 +3,9 @@
 Agents are autonomous entities that can perceive their environment, make decisions, and interact with other agents and the system they inhabit. They are the building blocks of Mango.jl, representing the individual entities or actors within a larger system.
 
 
-## 1. Agent Definition with @agent Macro
+## Agent Definition with @agent Macro
 
-To define an agent the `@agent` macro can be used. It simplifies the process of defining an agent struct and automatically adds necessary baseline fields. Here's how you can define an agent:
+To define an agent the [`@agent`](@ref) macro can be used. It simplifies the process of defining an agent struct and automatically adds necessary baseline fields. Here's how you can define an agent:
 
 ```julia
 using Mango
@@ -19,20 +19,11 @@ end
 my_agent = MyAgent("MyValue")
 ```
 
-The `@agent` macro adds the baseline fields listed in the table below. You can initialize the agent with exclusive fields like `my_own_field` in the example.
+The [`@agent`](@ref) macro adds some internal baseline fields to the struct. You can initialize the agent with exclusive fields like `my_own_field` in the example.
 
-| Field        | Description                                                    | Usable?                                                                           |
-|--------------|----------------------------------------------------------------|-----------------------------------------------------------------------------------|
-| aid          | The id of the agent                                            | Yes with aid(agent)!                                                               |
-| context      | Holds the reference to the container to send messages.         | Generally not recommended, use the convenience methods defined on the Agent type. |
-| scheduler    | The scheduler of the agent                                     | Generally not recommended, use the convenience methods defined on the Agent type. |
-| lock         | The agent lock to ensure only one message is handled per time. | Internal use only!                                                                |
-| role_handler | Contains the roles and handles their interactions              | Internal use only!                                                                |
+## Role Management
 
-
-## 2. Role Management
-
-Agents can have multiple roles associated with them. Roles can be added using the `add` function, allowing the agent to interact with its environment based on different roles. Here's how you can add roles to an agent:
+Agents can have multiple roles associated with them. Roles can be added using the [`add`](@ref) function, allowing the agent to interact with its environment based on different roles. Here's how you can add roles to an agent:
 
 ```julia
 using Mango
@@ -71,9 +62,9 @@ created_agent = add_agent_composed_of(your_container, RoleA(), RoleB(), RoleC())
 
 For more information on roles, take a look at [Role definition](@ref)
 
-## 3. Message Handling
+## Message Handling
 
-Agents and Roles can handle incoming messages through the `handle_message` function. By default, it does nothing, but you can override it to define message-specific behavior. You can also add custom message handlers for specific roles using the `subscribe_message` function. Here's how to handle messages:
+Agents and Roles can handle incoming messages through the [`handle_message`](@ref) function. By default, it does nothing, but you can override it to define message-specific behavior. You can also add custom message handlers for specific roles using the [`subscribe_message`](@ref) function. Here's how to handle messages:
 
 ```julia
 using Mango
@@ -96,7 +87,7 @@ function handle_message(role::MyRole, message::Any, meta::Any)
 end
 ```
 
-Besides the ability to handle messages, there also must be a possibility to send messages. This is implemented using the `send_message` function, defined on roles and agents.
+Besides the ability to handle messages, there also must be a possibility to send messages. This is implemented using the [`send_message`](@ref) function, defined on roles and agents.
 
 
 ```julia
@@ -117,10 +108,10 @@ send_message(agent, "Message", AgentAddress("receiver_id", "receiver_addr", "opt
[…]
