### 2024-08-21 c181d3a
remove importing of Mango.handle_message in examples
