### 2024-08-21 9916d71
update getting_started and tested example to the latest version
