+++ README.md
@@ -69,7 +69,6 @@ The following simple showcase demonstrates how you can define agents in Mango. J
 
 ```julia
 using Mango
-using Sockets: InetAddr, @ip_str
 
 # Create the container instances with TCP protocol
 container = create_tcp_container("127.0.0.1", 5555)
