### 2024-08-21 9815638
Modifying @role to use @kwdef instead of baseline constructor.
