### 2024-08-21 c844f75
Add @with_def as @kwdef alternative.
