### 2024-08-22 #7098 @jsagerOffis
Hi @Tortar @aurorarossi we have now merged some major changes onto the `joss_paper` and `development` branches, including:

* API documentation
* style-compliant docstrings
* `activate(containers) do ... end` syntax
* the installations instructions
* benchmarking information against mango-python
* examples in the README, paper, and tests are now using the new features and are functionally the same

All of this will make its way to `main` as release 0.4.0 but we may want to resolve some of the new issues before that.
