+++ src/util/datastructures_util.jl
@@ -32,13 +32,43 @@ function find_tuple(tuples::Vector, index::Int, value::Any)::Union{Tuple,Nothing
     return nothing
 end
 
+function get_types(type_curly_expr::Expr)
+    type_args = []
+    type_args_full = []
+    for i in 2:length(type_curly_expr.args)
+        type_arg = type_curly_expr.args[i]
+        push!(type_args_full, type_arg)
+        if typeof(type_arg) == Symbol
+            push!(type_args, type_arg)
+        else
+            push!(type_args, type_arg.args[1])
+        end
+    end
+    return type_args, type_args_full
+end
+
 macro with_def(struct_def)
     Base.remove_linenums!(struct_def)
 
     struct_head = struct_def.args[2]
     struct_name = struct_head
+    type_clause = nothing
+    type_args = []
+    type_args_full = []
     if typeof(struct_name) != Symbol
         struct_name = struct_head.args[1]
+        if typeof(struct_name) != Symbol
+            struct_head_sub = struct_name
+            struct_name = struct_head_sub.args[1]
+            if struct_head_sub.head == :curly
+                type_clause = struct_head_sub.args[2]
+                type_args, type_args_full = get_types(struct_head_sub)
+            end
+        end
+        if struct_head.head == :curly
+            type_clause = struct_head.args[2]
+            type_args, type_args_full = get_types(struct_head)
+        end
     end
 
     struct_fields = struct_def.args[3].args
@@ -66,19 +96,27 @@ macro with_def(struct_def)
         for (field, field_default) in new_struct_field_default
     ]
     args = new_struct_field
+    args_compl = length(type_args) == 0 ?
+                 Symbol(:new) :
+                 Expr(:curly, Symbol(:new), type_args...)
+
+    args_names_new = length(struct_field_names) == 0 ?
+                     Expr(:call, args_compl) :
+                     Expr(:call, args_compl, struct_field_names...)
+
+    call_clause = length(args) == 0 ? Expr(:call, Symbol(struct_name),
+        length(args_with_default) == 0 ? Expr(:parameters,) : Expr(:parameters, args_with_default...)
+    ) :
+                  Expr(:call, Symbol(struct_name),
+        length(args_with_default) == 0 ? Expr(:parameters,) : Expr(:parameters, args_with_default...),
+        args...
+    )
+    with_where_clause = length(type_args) == 0 ? call_clause : Expr(:where, call_clause, type_args_full...)
 
     default_constructor = Expr(:function,
-        length(args) == 0 ? Expr(:call, Symbol(struct_name),
-            length(args_with_default) == 0 ? Expr(:parameters,) : Expr(:parameters, args_with_default...)
-        ) :
-        Expr(:call, Symbol(struct_name),
-            length(args_with_default) == 0 ? Expr(:parameters,) : Expr(:parameters, args_with_default...),
-            args...
-        ),
-        Expr(:block, length(struct_field_names) == 0 ?
-                     Expr(:call, Symbol(:new)) :
-                     Expr(:call, Symbol(:new), struct_field_names...)
-        ))
+        with_where_clause,
+        Expr(:block, args_names_new)
+    )
 
     # Create the new struct definition
     new_struct_def = Expr(
