+++ test/datastructure_util_tests.jl
@@ -16,7 +16,18 @@ end
     i::Int = 0
 end
 
+@with_def struct ABC2 <: Int
+    abc::T
+    i::Int = 0
+end
+
 @testset "TestSimpleTypeParam" begin
     abc = ABC("H")
     @test abc.abc == "H"
+    @test abc.i == 0
+end
+@testset "TestTypeInheritanceWithDef" begin
+    abc = ABC("H")
+    @test abc.abc == "H"
+    @test abc.i == 0
 end
