### 2024-08-26 882ac7e
Add test for type inheritance only.
