+++ src/util/datastructures_util.jl
@@ -47,6 +47,22 @@ function get_types(type_curly_expr::Expr)
     return type_args, type_args_full
 end
 
+"""
+This macro can be applied to struct definitions and adds the possibility to define default values on field declaration. In contrast
+to @kwdef or @with_kw, this macro will create normal parameters for construction if no default value has been provided.
+
+# Example
+```julia
+@with_def struct ExampleStruct
+    abc::Int = 0
+    def::String
+end
+
+# can be constructed as follows
+es = ExampleStruct("mydefstring") # abc == 0
+es = ExampleStruct("mydefstring", abc=3) # abc == 3
+```
+"""
 macro with_def(struct_def)
     Base.remove_linenums!(struct_def)
 
