### 2024-08-26 f3aa83a
Added api docs for @with_def
