+++ docs/src/api.md
@@ -70,3 +70,13 @@ Modules = [Mango]
 Private = false
 Pages = ["util/encode_decode.jl"]
 ```
+
+# Misc
+
+In the following the built-in functions for encoding and decoding messages are listed.
+
+```@autodocs
+Modules = [Mango]
+Private = false
+Pages = ["util/datastructures_util.jl"]
+```
