### 2024-08-26 79029fd
Aligning Container and SimulationContainer, Access of roles via indexing, remove some debugging outputs.
