### 2024-08-26 a69b525
Fixing bug in add_agent_composed_of
