### 2024-08-29 #7098 @jsagerOffis
The intent of the example here was to highlight that (presumably) known functions from "non-role agents" function the same on roles which is why the function calls are just filled with dummy values and the agent is not registered to a container (because this is not a role-specific thing).

@aurorarossi Is it actually necessary for every individual code block to be executable on its own?
In my opinion, making sure each block here runs standalone would needlessly obscure the intended information of this part of the documentation with boilerplate code.
