### 2024-08-29 #7098 @jsagerOffis
We are currently trying to come up with an idea to make all code blocks executable while also saving a bit on boilerplate code for readability.

Also noted a mistake in the above example. second line should be MyRole of course:
```julia
agent = MyAgent("")
role = MyAgent("")
```
