### 2024-08-29 a8dd561
Adding some convenience functions and methods for agents, containers, topologies
