+++ src/world/topology.jl
@@ -1,4 +1,4 @@
-export complete_topology, star_topology, cycle_topology, graph_topology, per_node, add!, topology_neighbors, create_topology, add_node!, add_edge!, Topology
+export complete_topology, star_topology, cycle_topology, graph_topology, per_node, add!, topology_neighbors, create_topology, add_node!, add_edge!, Topology, modify_topolology, choose_agent, assign_agent, NORMAL, BROKEN, INACTIVE
 
 using MetaGraphsNext
 using Graphs
@@ -13,19 +13,23 @@ struct Topology
     graph::MetaGraph
 end
 
-@kwdef mutable struct TopologyService
-    neighbors::Vector{AgentAddress} = Vector()
+@enum State begin
+    NORMAL # normal neighbor
+    INACTIVE # neighbor link exists but link is not active (could be activated/used)
+    BROKEN # neighbor link exists but link is not usable (can not be activated)
 end
 
-function neighbors(service::TopologyService)
-    return service.neighbors
+@kwdef mutable struct TopologyService
+    state_to_neighbors::Dict{State,Vector{AgentAddress}} = Dict()
+end
+
+function neighbors(service::TopologyService, state::State=NORMAL)
+    return get(service.state_to_neighbors, state, Vector())
 end
 
 function _create_meta_graph_with(graph::AbstractGraph)
     vertices_description = [i => Node(id=i) for i in vertices(graph)]
-    edges_description = [
-        (e.src, e.dst) => nothing for e in edges(graph)
-    ]
+    edges_description = [(e.src, e.dst) => NORMAL for e in edges(graph)]
 
     return MetaGraph(graph, vertices_description, edges_description)
 end
@@ -75,13 +79,31 @@ end
 Add an edge to the topology from `node_id_from` to `node_id_to`. If `directed` is true
 a directed edge is added, otherwise an undirected edge is added.
 """
-function add_edge!(topology::Topology, node_id_from::Int, node_id_to::Int; directed::Bool=false)
+function add_edge!(topology::Topology, node_id_from::Int, node_id_to::Int, state::State=NORMAL; directed::Bool=false)
     if directed
-        topology.graph[node_id_from, node_id_to] = nothing
+        topology.graph[node_id_from, node_id_to] = state
     else
-        topology.graph[node_id_to, node_id_from] = nothing
-        topology.graph[node_id_from, node_id_to] = nothing
+        topology.graph[node_id_to, node_id_from] = state
+        topology.graph[node_id_from, node_id_to] = state
     end
+end
+
+"""
+    remove_edge!(topology::Topology, node_id_from::Int, node_id_to::Int)
+
+Remove the edge between `node_id_from` and `node_id_to`.
+"""
+function remove_edge!(topology::Topology, node_id_from::Int, node_id_to::Int)
+    return rem_edge!(topology.graph, node_id_from, node_id_to)
+end
+
+"""
+    remove_node!(topology::Topology, node_id::Int)
+
+Remove the node with the id `node_id`.
+"""
+function remove_node!(topology::Topology, node_id::Int)
+    return rem_vertex!(topology.graph, node_id)
 end
 
 """
@@ -89,10 +111,19 @@ end
 
 Add a node to the topology with a list (or a single) of agents attached.
 """
-function add_node!(topology::Topology, agents::Agent...)::Int
-    vid = nv(topology.graph) + 1
+function add_node!(topology::Topology, agents::Agent...; id::Union{Int,Nothing}=nothing)::Int
+    vid = isnothing(id) ? nv(topology.graph) + 1 : id
     topology.graph[vid] = Node(vid, [a for a in agents])
     return vid
+end
+
+"""
[…]
