+++ test/topology_tests.jl
@@ -44,11 +44,33 @@ end
     @test topology_neighbors(agents(container)[3]) == [address(agents(container)[1])]
 end
 
+@testset "TestModifyTopology" begin
+    container = create_tcp_container("127.0.0.1", 3333)
+    agent = nothing
+    topology = cycle_topology(4)
+
+    modify_topolology(topology) do topology
+        agent = register(container, TopologyAgent())
+        agent2 = register(container, TopologyAgent())
+        agent3 = register(container, TopologyAgent())
+        n1 = add_node!(topology, agent)
+        n2 = add_node!(topology, agent2)
+        n3 = add_node!(topology, agent3)
+        add_edge!(topology, n1, n2)
+        add_edge!(topology, n1, n3)
+    end
+
+    @test topology_neighbors(agent) == [address(agents(container)[2]),
+        address(agents(container)[3])]
+    @test topology_neighbors(agents(container)[2]) == [address(agents(container)[1])]
+    @test topology_neighbors(agents(container)[3]) == [address(agents(container)[1])]
+end
+
 @testset "TestCreateTopologyDirected" begin
     container = create_tcp_container("127.0.0.1", 3333)
     agent = nothing
 
-    create_topology() do topology
+    create_topology(directed=true) do topology
         agent = register(container, TopologyAgent())
         agent2 = register(container, TopologyAgent())
         agent3 = register(container, TopologyAgent())
@@ -114,4 +136,36 @@ end
     end
 
     @test length(topology_neighbors(tr)) == 4
+end
+
+@testset "TestTopologyChooseAgent" begin
+    topology = cycle_topology(4)
+    container = create_tcp_container("127.0.0.1", 3333)
+
+    choose_agent(topology) do node
+        return register(container, TopologyAgent())
+    end
+
+    @test length(topology_neighbors(agents(container)[1])) == 2
+    @test length(topology_neighbors(agents(container)[2])) == 2
+    @test length(topology_neighbors(agents(container)[3])) == 2
+    @test length(topology_neighbors(agents(container)[4])) == 2
+end
+
+@testset "TestTopologyAssignAgent" begin
+    topology = cycle_topology(4)
+    container = create_tcp_container("127.0.0.1", 3333)
+    register(container, TopologyAgent())
+    register(container, TopologyAgent())
+    register(container, TopologyAgent())
+    register(container, TopologyAgent())
+
+    assign_agent(topology, container) do agent, node
+        return aid(agent) == "agent" * string(node.id - 1)
+    end
+
+    @test length(topology_neighbors(agents(container)[1])) == 2
+    @test length(topology_neighbors(agents(container)[2])) == 2
+    @test length(topology_neighbors(agents(container)[3])) == 2
+    @test length(topology_neighbors(agents(container)[4])) == 2
 end
