+++ src/container/core.jl
@@ -3,6 +3,7 @@ export Container, notify_ready
 using Parameters
 using OrderedCollections
 using Base.Threads
+import Base.getindex
 
 # id key for the receiver
 RECEIVER_ID::String = "receiver_id"
@@ -233,3 +234,6 @@ function send_message(
     )
 end
 
+function getindex(container::Container, index::String)
+    return container.agents[index]
+end
