### 2024-08-29 f486270
set_edge_state + test.
