### 2024-08-30 #7098 @aurorarossi
> @aurorarossi Is it actually necessary for every individual code block to be executable on its own?
In my opinion, making sure each block here runs standalone would needlessly obscure the intended information of this part of the documentation with boilerplate code.

I run the above blocks as well. I agree that the single block should not run standalone, but with the above run blocks everything should work. Otherwise I think these examples are more confusing than helpful. But I think this is something standard at least for the Julia community.
