+++ src/express/api.jl
@@ -1,4 +1,4 @@
-export create_tcp_container, create_mqtt_container, GeneralAgent, add_agent_composed_of, activate
+export create_tcp_container, create_mqtt_container, GeneralAgent, add_agent_composed_of, agent_composed_of, activate, run_in_real_time, run_in_simulation, run_with_mqtt, run_with_tcp
 
 function _set_codec(container::Container, codec::Union{Nothing,Tuple{Function,Function}})
     if !isnothing(codec)
@@ -53,6 +53,27 @@ end
 end
 
 """
+    agent_composed_of(roles::Role...)
+
+Create an agent which is composed of the given roles `roles...`.
+
+The agent struct used is an empty struct, it is only used as a container
+for the roles. The created agent will be returned by the function.
+
+# Examples
+```julia
+agent = agent_composed_of(RoleA(), RoleB(), RoleC())
+```
+"""
+function agent_composed_of(roles::Role...)
+    agent = GeneralAgent()
+    for role in roles
+        add(agent, role)
+    end
+    return agent
+end
+
+"""
 	add_agent_composed_of(container, roles...)
 
 Create an agent which is composed of the given roles `roles...` and register the agent 
@@ -67,10 +88,7 @@ agent = add_agent_composed_of(your_container, RoleA(), RoleB(), RoleC())
 ```
 """
 function add_agent_composed_of(container::ContainerInterface, roles::Role...; suggested_aid::Union{Nothing,String}=nothing)
-    agent = GeneralAgent()
-    for role in roles
-        add(agent, role)
-    end
+    agent = agent_composed_of(roles...)
     register(container, agent, suggested_aid)
     return agent
 end
@@ -125,4 +143,113 @@ end
 
 function activate(runnable_simulation_code::Function, container::ContainerInterface)
     activate(runnable_simulation_code, [container])
+end
+
+"""
+    run_in_real_time(runnable::Function,
+    n_container::Int,
+    container_list_creator::Function,
+    agents::Union{Tuple,Agent}...)
+
+Let the agents run in containers (real time).
+
+Distributes the given `agents` to `n_container` (real time container) and execute the `runnable` 
+(takes the container list as argument) while the container are active to run. It is possible to 
+add supplementary information per agent as Tuple. For example `(Agent, :aid => "my_aid")`.
+The type of the containers are determined by the `container_list_creator` (n_container as argument, has
+to return a list of container with n_container entries).
+"""
+function run_in_real_time(runnable::Function,
+    n_container::Int,
+    container_list_creator::Function,
+    agents::Union{Tuple,Agent}...)
+
+    actual_number_container = n_container
+    if n_container < length(agents)
+        actual_number_container = length(agents)
+    end
+    container_list = container_list_creator(actual_number_container)
+    for (i, agent) in enumerate(agents)
+        container_id = ((i - 1) % actual_number_container) + 1
+        if isa(agent, Tuple)
+            actual_agent = agent[1]
+            agent_params::Dict = Dict(agent[2])
+            register(container_list[container_id], actual_agent, get(agent_params, :aid, nothing),
+                topics=get(agent_params, :topics, []))
+        else
+            register(container_list[container_id], agent)
+        end
+    end
+    activate(container_list) do
+        runnable(container_list)
+    end
+end
+
+"""
+    run_with_tcp(runnable::Function,
+    n_container::Int,
+    agents::Union{Tuple,Agent}...;
+    host::String="127.0.0.1",
+    start_port::Int=5555,
+    codec::Union{Nothing,Tuple{Function,Function}}=(encode, decode))
+
+Let the agents run in tcp containers (real time).
+
[…]
