### 2024-09-03 dc73d9c
Overhaul of doc examples. Enable testing of doc examples and including some fixes.
