+++ src/express/api.jl
@@ -1,4 +1,4 @@
-export create_tcp_container, create_mqtt_container, GeneralAgent, add_agent_composed_of, agent_composed_of, activate, run_in_real_time, run_in_simulation, run_with_mqtt, run_with_tcp
+export create_tcp_container, create_mqtt_container, GeneralAgent, add_agent_composed_of, agent_composed_of, activate, run_in_real_time, run_in_simulation, run_with_mqtt, run_with_tcp, PrintingAgent
 
 function _set_codec(container::Container, codec::Union{Nothing,Tuple{Function,Function}})
     if !isnothing(codec)
@@ -20,7 +20,7 @@ can also provide a codec as tuple of functions (first encode, second decode, see
 agent = create_mqtt_container("127.0.0.1", 5555, "MyClient")
 ```
 """
-function create_tcp_container(host::String, port::Int; codec::Union{Nothing,Tuple{Function,Function}}=nothing)
+function create_tcp_container(host::String="127.0.0.1", port::Int=5555; codec::Union{Nothing,Tuple{Function,Function}}=nothing)
     container = Container()
     container.protocol = TCPProtocol(address=InetAddr(host, port))
     _set_codec(container, codec)
@@ -42,7 +42,7 @@ Optionally you can also provide a codec as tuple of functions (first encode, sec
 agent = create_mqtt_container("127.0.0.1", 5555, "MyClient")
 ```
 """
-function create_mqtt_container(host::String, port::Int, client_id::String; codec::Union{Nothing,Tuple{Function,Function}}=nothing)
+function create_mqtt_container(host::String="127.0.0.1", port::Int=1883, client_id::String="default"; codec::Union{Nothing,Tuple{Function,Function}}=nothing)
     container = Container()
     container.protocol = MQTTProtocol(client_id, InetAddr(host, port))
     _set_codec(container, codec)
@@ -53,20 +53,20 @@ end
 end
 
 """
-    agent_composed_of(roles::Role...)
+    agent_composed_of(roles::Role...; base_agent::Union{Nothing,Agent}=nothing)
 
 Create an agent which is composed of the given roles `roles...`.
 
-The agent struct used is an empty struct, it is only used as a container
-for the roles. The created agent will be returned by the function.
-
+If no `base_agent` is provided the agent struct used is an empty struct, which is 
+only used as a container for the roles. The agent will be returned by the function.
+    
 # Examples
 ```julia
 agent = agent_composed_of(RoleA(), RoleB(), RoleC())
 ```
 """
-function agent_composed_of(roles::Role...)
-    agent = GeneralAgent()
+function agent_composed_of(roles::Role...; base_agent::Union{Nothing,Agent}=nothing)
+    agent = isnothing(base_agent) ? GeneralAgent() : base_agent
     for role in roles
         add(agent, role)
     end
@@ -74,21 +74,21 @@ function agent_composed_of(roles::Role...)
 end
 
 """
-	add_agent_composed_of(container, roles...)
+    add_agent_composed_of(container, roles::Role...; suggested_aid::Union{Nothing,String}=nothing, base_agent::Union{Nothing,Agent}=nothing)
 
 Create an agent which is composed of the given roles `roles...` and register the agent 
 to the `container`. 
 
-The agent struct used is an empty struct, it is only used as a container
-for the roles. The created agent will be returned by the function.
+If no `base_agent` is provided the agent struct used is an empty struct, which is 
+only used as a container for the roles. The agent will be returned by the function.
 
 # Examples
 ```julia
 agent = add_agent_composed_of(your_container, RoleA(), RoleB(), RoleC())
 ```
 """
-function add_agent_composed_of(container::ContainerInterface, roles::Role...; suggested_aid::Union{Nothing,String}=nothing)
[…]
