+++ src/container/core.jl
@@ -242,6 +242,6 @@ function Base.getindex(container::Container, index::String)
     return container.agents[index]
 end
 
-function getindex(container::Container, index::Int)
+function Base.getindex(container::Container, index::Int)
     return agents(container)[index]
 end
