+++ docs/src/topology.md
@@ -8,11 +8,16 @@ As it can be pretty clunky to create every neighborhood-list manually, Mango.jl 
 
 First, there are several pre-defined topologies. It is also possible to use an arbitrary Graphs.jl graph. After the creation of the topology, the agents need to be added to the topology. This can be done with `per_node(topology) do node ... end`. In the do-block it is possible to add agents to nodes, the do-block will be executed per vertex of your graph. 
 
-```julia
-topology = complete_topology() # fully connected 
-topology = star_topology() # star
-topology = cycle_topology() # cycle
-topology = graph_topology(your_graph) # based on arbitrary Graphs.jl AbstractGraph
+```@example
+using Mango, Graphs
+
+@agent struct MyAgent end
+
+topology = star_topology(3) # star
+topology = cycle_topology(3) # cycle
+topology = complete_topology(3) # fully connected 
+topology = graph_topology(complete_digraph(3)) # based on arbitrary Graphs.jl AbstractGraph
+
 per_node(topology) do node
     add!(node, MyAgent())
 end
@@ -20,19 +25,27 @@ end
 
 However, often this approach is not feasible, because you create a specific agent system with agents which need to be linked in a very specific way, such that it is not possible to assign the same agent type to every node. For this reason you can define the topology manually:
 
-```julia
+```@example
+using Mango 
+
+@agent struct TopologyAgent end
+container = Container()
+
 topology = create_topology() do topology
-    agent = register(container, TopologyAgent())
+    agent0 = register(container, TopologyAgent())
+    agent1 = register(container, TopologyAgent())
     agent2 = register(container, TopologyAgent())
-    agent3 = register(container, TopologyAgent())
-    n1 = add_node!(topology, agent)
-    n2 = add_node!(topology, agent2)
-    n3 = add_node!(topology, agent3)
+    n1 = add_node!(topology, agent0)
+    n2 = add_node!(topology, agent1)
+    n3 = add_node!(topology, agent2)
     add_edge!(topology, n1, n2)
     add_edge!(topology, n1, n3)
 end
+
+# neighbors of `agent`
+topology_neighbors(container[1])
 ```
 
 # Using the topology
 
-At this point we know how to create topologies and how to populate them. To actually use them, the function `topology_neighbors(agent_or_role)` exists. The function returns a vector of AgentAddress objects, which represent all other agents in the neighborhood of `agent`.
+At this point we know how to create topologies and how to populate them. To actually use them, the function [`topology_neighbors`](@ref) exists. The function returns a vector of AgentAddress objects, which represent all other agents in the neighborhood of `agent`.
