### 2024-09-04 9106ca3
More illustrative scheduling example.
