### 2024-09-04 cf1b527
Adding express tests and using multiple dispatch instead of type union.
