### 2024-09-04 #7098 @jsagerOffis
@aurorarossi  Just merged the new changes into `development` and `joss_paper`.
All examples should now be properly executable.
