### 2024-09-08 e1b25e4
Add global events.
