+++ versions.js
@@ -2,6 +2,7 @@ var DOC_VERSIONS = [
   "stable",
   "v0.3",
   "v0.2",
+  "dev",
 ];
 var DOCUMENTER_NEWEST = "v0.3.0";
 var DOCUMENTER_STABLE = "stable";
