+++ .github/workflows/test-mango.yml
@@ -6,7 +6,7 @@ on:
       - main
       - development
   pull_request:
-    types: [opened, reopened]
+    types: [opened, synchronize, reopened]
 
 
 # needed to allow julia-actions/cache to delete caches
