### 2024-09-16 #7098 @aurorarossi
Ok I checked the examples and they work. I also made a pull request to fix some typos and improve the flow of some sentences.
