### 2024-09-27 #7098 @Tortar
Hy @jbytecode sorry for the delay, I'm happy with all of this now. Great work!

Just one small final detail: given that there is some code in the paper, I think it could be useful to mention the version of Mango.jl with which it has been executed for easier future reproduction
