### 2024-09-27 #7098 @jsagerOffis
@jbytecode I just pushed a line giving the version to the paper. The changes made during the review will be their own new version and also receive a tagged branch.
