### 2024-09-30 #7098 @jbytecode
@jsagerOffis - The default branch of the paper is set to `joss_paper` in this thread. However, the software repository doesn't have a branch of this name. Could you please clarify this issue? I need to fork the repo for some minor changes and I don't want to edit wrong files in wrong places. Thank you in advance.
