### 2024-09-30 #7098 @jsagerOffis
@jbytecode
* Double check authors and affiliations (including ORCIDs) - done
* Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. - done, v0.4.0
* Archive the release on Zenodo/figshare/etc and post the DOI here. - done, DOI: 10.5281/zenodo.13860452
* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. - done
* Make sure that the license listed for the archive is the same as the software license. - done
