### 2024-09-30 #7098 @jbytecode
@jsagerOffis - The author names, their DOIs, and the **title of the manuscript** should match that the ones in the Zenodo archive. Please fix this and then ping me again.
