### 2024-10-01 1e6ade2
update bibtex
