+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{schrage:2024,
-  title={mango: A modular python-based agent simulation framework},
+  title={mango: A modular {P}ython-based agent simulation framework},
   author={Schrage, Rico and Sager, Jens and H{\"o}rding, Jan Philipp and Holly, Stefanie},
   journal={SoftwareX},
   volume={27},
@@ -127,7 +127,7 @@
     and Dancy, Christopher
     and Hyder, Ayaz
     and Hussain, Muhammad",
-    title="Utilizing Python for Agent-Based Modeling: The Mesa Framework",
+    title="Utilizing {P}ython for Agent-Based Modeling: The Mesa Framework",
     booktitle="Social, Cultural, and Behavioral Modeling",
     year="2020",
     publisher="Springer International Publishing",
