### 2024-10-01 #7098 @jbytecode
@jsagerOffis - I've just sent a pull request about minor fixes in both the manuscript and bibtex. Please review the PR and tell me if there is something you are not agreed with. If it looks okay to you, please merge. Thank you in advance.

https://github.com/OFFIS-DAI/Mango.jl/pull/72
