### 2024-10-01 #7098 @jsagerOffis
@jbytecode I just merged it to development and also merged those changes back into the joss_paper branch so no worries.

I will also move the final paper state up to the main branch and move the release tag to that commit once its all final.
