### 2024-10-01 #7098 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
