### 2024-10-01 3a6da6b
small changes in paper.md
