+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 # Summary
 Multi-agent simulations are inherently complex, making them difficult to implement, maintain, and optimize.
-An agent, as defined by [@russel:2010], is software that perceives its environment through sensors and acts upon it using actuators.
+An agent, as defined by @russel:2010, is software that perceives its environment through sensors and acts upon it using actuators.
 `Mango.jl` is a simulation framework for multi-agent systems implemented in Julia [@julia:2017].
 It enables quick implementations of multiple communicating agents, either spanning multiple devices or in a single local environment.
 
@@ -50,7 +50,7 @@ This is useful for simulations, where simulated time should run much faster than
 
 
 # Statement of need
-Applications of multi-agent systems can be found in various fields, such as in distributed optimization [@yang:2019], reinforcement learning [@gronauer:2022], robotics [@chen:2019] and more.
+Applications of multi-agent systems can be found in various fields, such as in distributed optimization [@yang:2019], reinforcement learning [@gronauer:2022], robotics [@chen:2019], and more.
 Many of these systems are highly complex and feature heterogeneous and interacting actors.
 This makes them inherently difficult to model and develop.
 Therefore, a structured development framework to support this process is a valuable asset.
@@ -59,7 +59,7 @@ While `Mango.jl` is a general purpose multi-agent framework, we will focus on en
 
 Many of the ideas for `Mango.jl` are based on the existing Python framework `mango` [@schrage:2024]. 
 The main reason for this Julia-based version is to allow better focus on simulation performance, enabling larger scales of multi-agent simulations.
-This is especially relevant in the energy domain, where an increasing amount of energy resources (e.g. batteries and PV-generators) have distributed ownership, competing goals and contribute to the same power grid.
+This is especially relevant in the energy domain, where an increasing amount of energy resources (e.g., batteries and PV-generators) have distributed ownership, competing goals and contribute to the same power grid.
 Large scale multi-agent simulations allow researchers to study the behavior of these participants in energy markets and grid simulations.
 
 The Python version of `mango` has already been successfully applied to various research areas in the energy domain, including coalition formation in multi-energy networks [@schrage:2023], distributed market participation of battery storage units [@tiemann:2022], distributed black start [@stark:2021], and investigating the impact of communication topologies on distributed optimization heuristics [@holly:2021].
@@ -69,8 +69,8 @@ New Julia-based projects using `Mango.jl` are in active development.
 To our knowledge, there is no Julia-based multi-agent framework with a focus on agent communication and distributed operation like `Mango.jl`.
 
 `Agents.jl` [@agents:2022] is a multi-agent framework for modeling agent interactions in a defined space to observe emergent properties like in animal flocking behavior or the spreading of diseases. 
-This puts it in line with frameworks like mesa [@mesa:2020] or NetLogo [@netlogo:2004].
-These have a different scope than `Mango.jl` which is more focused on agent communication and internal agent logic for software applications.
+This puts it in line with frameworks like mesa [@mesa:2020] and NetLogo [@netlogo:2004].
+These have a different scope than `Mango.jl`, which is more focused on agent communication and internal agent logic for software applications.
 
 JADE [@JADE:2001] and JIAC [@jiac:2013] are Java frameworks of similar scope but are not actively developed anymore. 
 JACK [@jack:2005] provides a language and tools to implement communicating agents but is discontinued and proprietary.
@@ -89,14 +89,14 @@ All workloads in the agents are entirely simulated by static delays.
 Thus, the benchmarks assume that workloads in Python and Julia are identical.
 
 The main advantage of `Mango.jl` is in the ease of parallelization.
[…]
