### 2024-10-01 #7098 @danielskatz
👋 @jsagerOffis - As the AEiC who will complete the processing of this paper, I've proofread it and suggest the changes in https://github.com/OFFIS-DAI/Mango.jl/pull/73. Please merge this, or let me know what you disagree with, then we can continue the acceptance and publication process.
